Split the node allocator and COW string into separate files.
